Dynamic parser (#5)
* Initial commit

* added files

* package framework

* main switch case added

* Revert "main switch case added"

This reverts commit 4f201061e007886f90e2386d57ec63d346de16ed.

* Revert "Revert "main switch case added""

This reverts commit 7e40d4bf2ceeadc541dda4a619731272809c9916.

* csvparser.js --- takes the raw data from csv and changes it into json
form, prints it into the intermediates folder line by line for each object
jsontagger.js --- takes the intermediate json strings line by line,
reformats the data, and saves it line by line.
function parse(data) --- reformats a known number of PDO's into the
database model

* made sure that files are printed with newlines between objects

parser.js --- contains the actual parsing functions for JSON objects as an
exportable module. May have to change the input to this function

* changed filetype to .json instead of .csv for the output

* ignore intellij files

* refactor: added subfolders to organize project
-csv/ contains scripts to convert the example input and write it to file
-serial/ contains json reformatter the feeds into the database

* Documentation

* serial port dependency

* serial reads data from arduino

* refactor parser

* jsontagger.js is now unnecessary, csvparser.js has been hookedup to the
parser as a stream

parser.js --- added parsing functionality for most of the important data
sources. exports itself as a Transform stream where data can be piped to
places.

also working on mocha tests for the parser so that we can do CI

* Frontend placeholder is made

* jsontagger.js is now unnecessary, csvparser.js has been hookedup to the
parser as a stream

parser.js --- added parsing functionality for most of the important data
sources. exports itself as a Transform stream where data can be piped to
places.

also working on mocha tests for the parser so that we can do CI

* Added mocha for testing framework

* index.js --- using the serial and parser modules, the server backend is
now complete, serial data is directly piped over into the parser, line by
line. Database and Serial still need to be implemented as read/write
streams respectively, but they will be easy to do so.

parser.js --- still trying to figure out the bit shifts to get the pack
maximum voltage

* index.js --- using the serial and parser modules, the server backend is
now complete, serial data is directly piped over into the parser, line by
line. Database and Serial still need to be implemented as read/write
streams respectively, but they will be easy to do so.

parser.js --- still trying to figure out the bit shifts to get the pack
maximum voltage

* added socket.io basic functionality to the front end and it now displays
the parsed data as paragraphs.

added testserver.js to send data from the outputted json

* reduced size of parser

* parser works except for the throttle and brake values, the csv file is
weird

* rename throttle and brake states to flag

* rename throttle and brake states to flag

* moved to class, this doesn't work yet

* its so broken

* serial.js --- have to split up the file, using callback functions requires
the var self=this magic variable. pretty much the only thing that wasn't
working, just split up the monolothic finding of the arduino from the
actual logic to send data. Furthermore, serialport has no pause or resume
functions, so the thing just reads.

serial_test/serial_test.ino --- added a line number to each data sent.

* ---testserver.js - reading now from the raw input csv, added a delay by
setTimeout() where the parser is paused and then resumed. Or else the
client gets all the data only after all of it has been parsed.

---created a new testparser to parse the old csv, whereas the new parser
parses the actual byte data and can be configured to be dynamic

---the parser now parses the bitmask of the flags into an array of true,
false statements

* created a demos directory that will be used to test on the csv data, the
top level directories contain the actual implementation. The serial also
attempts to reconnect to the arduino automatically

* Revert "created a demos directory that will be used to test on the csv data, the"

This reverts commit f54c65ea0ab442826147d6c2a9b42eb5e6aae553.

* Fixed the colossal deletion

* Fixed the colossal deletion

* database stuff

* added dbstream class

* database stream update

* delete unused files

* database stuff

* added dbstream class

* added dbstream class

* pushed up the random stuff that shouldn't be merged

* delete the last remnants of the structure

* First merge of the official database

* fixed dbStream only saving the first file
---dbStream.js- make sure callback() is called to signal nodejs that the chunk has been processed

---index.js in test now reflects the newer database testing methods

removed demo data

returned package.json to it's original state

Update README.md

need to install node-gyp for project to work

* fixed dbStream only saving the first file
---dbStream.js- make sure callback() is called to signal nodejs that the chunk has been processed

---index.js in test now reflects the newer database testing methods

* fix extra dependency async

* added a new api to get db requests from the server

* database page now displays
next step is to interface with the api and work on the drop down list and
the table population to look at log data

bms should have a BMSController

* added restful api to start and stop datalogging

--dbStream.js made sure that the stream disconnects from the database when
it is unpiped.

* fixed dbStream only saving the first file
---dbStream.js- make sure callback() is called to signal nodejs that the chunk has been processed

---index.js in test now reflects the newer database testing methods

* fix extra dependency async

* added a new api to get db requests from the server

* database page now displays
next step is to interface with the api and work on the drop down list and
the table population to look at log data

bms should have a BMSController

* bug fix server-side api, serial, and parser

--app.js, moved routing into routes.js
--serial.js use byte delimiter instead of newline parser
--parser.js only outputs strings of javascript objects
--serial_test.ino outputs dummy data now to test the server

* bug fix server-side api, serial, and parser

--app.js, moved routing into routes.js
--serial.js use byte delimiter instead of newline parser
--parser.js only outputs strings of javascript objects
--serial_test.ino outputs dummy data now to test the server
--controller.js uses the right function to send data over now, not the
shell functions, but the actual api

* app.js now splits up CAN messages by socket event

* Add /name route

Also remove extraneous _open definition

app.js pass database as a variable

* index database on Timestamp and CAN

* index database on Timestamp and CAN

--parser.js also State should be state in the database

* added dynamic parsing capability
--defaults.json - lists the default parsing setup
--parse_descriptor access the parsing descriptions in the database, also
loads the default into the database
--dynamicParser.js core of the parser
--test.js tests the parser with some sample data

* deep copy descriptor in array to parser

* save new definition when not found

* made the parser async and non-blocking

* made the parser async and non-blocking

--returned back to using fcall